requirements.txt
appdirs==1.4.0
chardet2==2.0.3
packaging==16.8
pdb==0.1
pdfminer2==20151206
pymining==0.2
pyparsing==2.1.10
python-gnupg==0.4.0
PyYAML==3.12
six==1.10.0
